+++ include/lin_sys.h
@@ -14,6 +14,7 @@ extern "C" {
 // Include solvers
 #include "suitesparse_ldl.h"
 
+
 #ifndef EMBEDDED
 // Initialize linear system solver structure
 // NB: Only the upper triangular part of P is stuffed!
@@ -29,30 +30,8 @@ LinSysSolver * init_linsys_solver(const csc * P, const csc * A,
                 const OSQPSettings *settings, c_int polish);
 
 
-// Free linear system solver
-// void free_linsys_solver(LinSysSolver * s);
-
-
-// // Initialize private variable for solver
-// // NB: Only the upper triangular part of P is stuffed!
-// Priv *init_priv(const csc * P, const csc * A,
-//                 const OSQPSettings *settings, c_int polish);
-//
-// // Free LDL Factorization structure
-// void free_priv(Priv *p);
-
 #endif
 
-
-// #if EMBEDDED != 1
-// // Update private structure with new P and A
-// c_int update_priv(Priv * p, const csc *P, const csc *A,
-//                   const OSQPWorkspace * work, const OSQPSettings *settings);
-// #endif
-
-// [> solves Ax = b for x, and stores result in b <]
-// c_int solve_lin_sys(const OSQPSettings *settings, Priv *p, c_float *b);
-//
 
 #ifdef __cplusplus
 }