+++ interfaces/python/extension/include/osqpworkspacepy.h
@@ -199,6 +199,18 @@
 
 
 static PyObject *OSQP_get_workspace(OSQP *self){
+
+    // Check if linear systems solver is SUITESPARSE_LDL
+    if(!self->workspace){
+        PyErr_SetString(PyExc_ValueError, "Solver is uninitialized.  No data have been configured.");
+        return (PyObject *) NULL;
+    }
+
+    if(self->workspace->linsys_solver->type != SUITESPARSE_LDL){
+        PyErr_SetString(PyExc_ValueError, "OSQP setup was not performed using SuiteSparse LDL! Run setup with linsys_solver as SuiteSparse LDL");
+        return (PyObject *) NULL;
+    }
+
      PyObject *data_py = OSQP_get_data(self);
      PyObject *linsys_solver_py = OSQP_get_linsys_solver(self);
      PyObject *scaling_py = OSQP_get_scaling(self);